+++ types/effect/node/module.f.ts
@@ -1,17 +1,44 @@
 import type { Vec } from "../../bit_vec/module.f.ts"
 import type { Result } from "../../result/module.f.ts"
-import type { Effect, ToAsyncOperationMap } from "../module.f.ts"
+import { type Do, type Effect, type ToAsyncOperationMap, type Operations, do_ } from "../module.f.ts"
 
 export type IoResult<T> = Result<T, unknown>
 
+export type MakeDirectoryOptions = { readonly recursive: true }
+
+export type MkdirParam = readonly[string, MakeDirectoryOptions?]
+
+export type WriteFileParam = readonly[string, Vec]
+
 export type FileOperations = {
-    readonly readFile: readonly[string, IoResult<Vec>]
-    readonly writeFile: readonly[readonly[string, Vec], IoResult<void>]
+    readonly readFile: readonly [string, IoResult<Vec>]
+    readonly writeFile: readonly [WriteFileParam, IoResult<void>]
+    readonly mkdir: readonly [MkdirParam, IoResult<void>]
 }
 
+export const readFile =
+    (path: string) =>
+    <O extends FileOperations, T>(f: (r: IoResult<Vec>) => Effect<O, T>): Do<O, T> =>
+    do_('readFile', path, f)
+
+export const writeFile =
+    (...p: WriteFileParam) =>
+    <O extends FileOperations, T>(f: (r: IoResult<void>) => Effect<O, T>): Do<O, T> =>
+    do_('writeFile', p, f)
+
+export const mkdir =
+    (...p: MkdirParam) =>
+    <O extends FileOperations, T>(f: (r: IoResult<void>) => Effect<O, T>): Do<O, T> =>
+    do_('mkdir', p, f)
+
 export type ConsoleOperations = {
-    readonly log: readonly[string, void]
+    readonly log: readonly [string, void]
 }
+
+export const log =
+    (msg: string) =>
+    <O extends ConsoleOperations, T>(f: () => Effect<O, T>): Do<O, T> =>
+    do_('log', msg, f)
 
 export type NodeOperations =
     & ConsoleOperations